backend/requirements.txt
fastapi
uvicorn[standard]
pytest
bcrypt==4.0.1
passlib==1.7.4
python-jose
lxml
python-multipart
httpx
firebase-admin
google-generativeai
openpyxl
xlrd
